Batch plugin improvements
https://github.com/aglushkov/serega/issues/94

- Show error with tips what goes wrong when batch plugin not added to root serializer
- move included modules to separate subfolder
